refactor: group health-check utils by domain (fileRequirements/modRequirements/shared) and use @/ imports (LAZ-590)
