+++ src/renderer/src/extensions/health_check/utils/fileRequirements/fileDependencyCache.ts
