+++ src/renderer/src/extensions/health_check/utils/shared/severityStyles.ts
